requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
boto3==1.12.0
botocore==1.15.0
Django==3.0.3
django-cleanup==4.0.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-graphql-playground==1.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
docutils==0.15.2
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.1
gunicorn==20.0.4
jmespath==0.9.4
Pillow==7.0.0
promise==2.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
